requirements.txt
Flask~=3.0.3
SQLAlchemy~=2.0.35
Flask-SQLAlchemy
Flask-Migrate
Jinja2
python-dotenv~=1.0.1
Werkzeug~=3.0.4
OpenAI~=1.51.2
gradio~=5.0.2

WTForms~=3.1.2
alembic~=1.13.3
tenacity~=9.0.0